requirements.txt
djangorestframework==3.14.0
PyYAML==6.0
Pygments==2.14.0
asgiref==3.6.0
pytz==2023.2
setuptools==65.5.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
wheel==0.38.4
# Django	4.1.7
# django-filter	23.1	23.1
# pygments-pprint-sql	0.1.0	0.1.0
# pip	22.3.1	23.0.1